Add files via upload
A phase-field model for brittle fracture is implemented in the commercial finite element software Abaqus by means of UEL subroutines. The phase-field method considerably reduces the implementation complexity for fracture problems as it removes the need for numerical tracking of discontinuities in the displacement field that is characteristic of discrete crack methods.